elang/api: Use |size_t| instead of |int| for offset parameter in |MachineCodeBuilder| API.

+++ elang/api/machine_code_builder.h
@@ -24,18 +24,18 @@ class MachineCodeBuilder {
   MachineCodeBuilder();
   virtual ~MachineCodeBuilder();
 
-  virtual void EmitCode(const uint8_t* codes, int code_size) = 0;
+  virtual void EmitCode(const uint8_t* codes, size_t code_size) = 0;
   virtual void FinishCode() = 0;
-  virtual void PrepareCode(int code_size) = 0;
-  virtual void SetCallSite(int offset, base::StringPiece16 string) = 0;
-  virtual void SetCodeOffset(int offset, int target_offset) = 0;
-  virtual void SetFloat32(int offset, float32_t float32) = 0;
-  virtual void SetFloat64(int offset, float64_t float64) = 0;
-  virtual void SetInt32(int offset, int32_t int32) = 0;
-  virtual void SetInt64(int offset, int64_t int64) = 0;
-  virtual void SetSourceCodeLocation(int offset,
+  virtual void PrepareCode(size_t code_size) = 0;
+  virtual void SetCallSite(size_t offset, base::StringPiece16 string) = 0;
+  virtual void SetCodeOffset(size_t offset, size_t target_offset) = 0;
+  virtual void SetFloat32(size_t offset, float32_t float32) = 0;
+  virtual void SetFloat64(size_t offset, float64_t float64) = 0;
+  virtual void SetInt32(size_t offset, int32_t int32) = 0;
+  virtual void SetInt64(size_t offset, int64_t int64) = 0;
+  virtual void SetSourceCodeLocation(size_t offset,
                                      SourceCodeLocation location) = 0;
-  virtual void SetString(int offset, base::StringPiece16 string) = 0;
+  virtual void SetString(size_t offset, base::StringPiece16 string) = 0;
 
  private:
   DISALLOW_COPY_AND_ASSIGN(MachineCodeBuilder);
